idm_23.2.1_release (#161)

+++ FMWKubernetesMAA/OracleEnterpriseDeploymentAutomation/OracleIdentityManagement/utils/delete_elk.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-# Copyright (c) 2022, Oracle and/or its affiliates.
+# Copyright (c) 2022, 2023, Oracle and/or its affiliates.
 # Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 #
 # This is an example of a script which will delete an Elastic Search deployment
@@ -7,12 +7,38 @@
 # Dependencies: ../common/functions.sh
 #               ../responsefile/idm.rsp
 #
-# Usage: delete_elk.sh
+# Usage: delete_elk.sh  [-r responsefile]
 #
-MYDIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+SCRIPTDIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+SCRIPTDIR=$SCRIPTDIR/..
 
-. $MYDIR/../common/functions.sh
+while getopts 'r:p:' OPTION
+do
+  case "$OPTION" in
+    r)
+      RSPFILE=$SCRIPTDIR/responsefile/$OPTARG
+     ;;
+    p)
+      PWDFILE=$SCRIPTDIR/responsefile/$OPTARG
+     ;;
+    ?)
+      echo "script usage: $(basename $0) [-r responsefile] " >&2
+      exit 1
+      ;;
+   esac
+done
+
+
+RSPFILE=${RSPFILE=$SCRIPTDIR/responsefile/idm.rsp}
+
 . $RSPFILE
+if [ $? -gt 0 ]
+then
+    echo "Responsefile : $RSPFILE does not exist."
+    exit 1
+fi
+
+. $SCRIPTDIR/common/functions.sh
 
 WORKDIR=$LOCAL_WORKDIR/ELK
 
@@ -46,6 +72,9 @@ echo "Delete Namespace $ELKNS"
 
 kubectl delete namespace $ELKNS  >> $LOG 2>&1
 
-rm  -rf $WORKDIR/logs $LOCAL_WORKDIR/elk_installed
+if [ ! "$WORKDIR" = "" ] && [ ! "$LOCAL_WORKDIR" = "" ]
+then
+  rm  -rf $WORKDIR/logs $LOCAL_WORKDIR/elk_installed
+fi
 FINISH_TIME=`date +%s`
 print_time TOTAL "Delete ELK " $START_TIME $FINISH_TIME
